# openapi.yaml
openapi: 3.0.0
info:
  title: Financial Advisor Backend API
  description: Backend service for the financial advisor frontend.
  version: "1.0"
servers:
  - url: http://localhost:8083
    description: Development server
paths:
  /agent/get-sr-list:
    post:
      summary: Fetch SR data
      description: Retrieves the SR data by identity.
      operationId: getServiceList
      parameters:
        - name: authorization
          in: header
          required: true
          description: Bearer token for authorization.
          schema:
            type: string
            example: "Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiIs..."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - pageNumber
                - pageCount
                - reqType
                - reqStatus
              properties:
                pageNumber:
                  type: integer
                  description: The page number for pagination.
                  example: 1
                pageCount:
                  type: integer
                  description: The number of items per page.
                  example: 100
                reqType:
                  type: array
                  items:
                    type: string
                  description: List of subcategory IDs to filter by.
                  example:
                    [
                      "2fc7ee5bd20544feb3d8a7db247ff30d",
                      "8b1a35a3ac7a4a6f863c1fb34de40845",
                    ]
                reqStatus:
                  type: array
                  items:
                    type: string
                  description: List of status IDs to filter by.
                  example: ["d0cc0947a9f34d099e66048dc64c1740"]
      responses:
        "200":
          description: Hierarchy data retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    description: HTTP status code
                    example: 200
                  message:
                    type: string
                    description: Status message
                    example: "SR Data retrieved successfully"
                  data:
                    type: array
                    description: List of meta data objects
                    items:
                      type: object
                      properties:
                        ticket:
                          type: string
                          description: Unique identifier for the SR
                          example: "4678e1bb1f2d414393a85dfbe0c85fff"
                        created_date:
                          type: string
                          description: Date of the request raised
                          example: "19 Nov 2024"
                        sr_closed_time:
                          type: string
                          description: Date of the request closed
                          example: "19 Nov 2024"
                        categoryname:
                          type: string
                          description: Category of the SR
                          example: "Nominee"
                        sub_category_name:
                          type: string
                          description: Sub-Category of the SR
                          example: "Nominee"
                        sr_status:
                          type: string
                          description: Status of the SR
                          example: "Approved"
        "401":
          description: Unauthorized - Invalid token
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unauthorized access."
...
  /agent/get-sr-details:
    get:
      summary: Get SR Details
      description: Retrieves the SR Details.
      operationId: getServiceDetails
      parameters:
        - name: authorization
          in: header
          required: true
          description: Bearer token for authorization.
          schema:
            type: string
            example: "Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiIs..."
        - name: sr_number
          in: query
          required: true
          description: The service request (SR) number to filter by.
          schema:
            type: string
            example: "9711"
      responses:
        "200":
          description: SR Data retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: integer
                    description: HTTP status code
                    example: 200
                  message:
                    type: string
                    description: Status message
                    example: "SR Data retrieved successfully"
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        sr_number:
                          type: string
                          description: Unique Service Request number
                          example: "2844"
                        created_date:
                          type: string
                          format: date
                          description: The date the SR was created
                          example: "29 Nov 2024"
                        closed_date:
                          type: string
                          format: date
                          description: The date the SR was closed
                          example: "29 Nov 2024"
                        values:
                          type: object
                          properties:
                            identity_nominee:
                              type: string
                              description: Nominee identifier
                              example: "f0df10820e5046cc9f04aa29ade62eb4"
                            oldValues:
                              type: object
                              properties:
                                title:
                                  type: string
                                  description: The old title of the nominee
                                  example: "Mr"
                                dob:
                                  type: string
                                  description: The old date of birth of the nominee
                                  example: ""
                                fullname:
                                  type: string
                                  description: The old full name of the nominee
                                  example: ""
                                relationship:
                                  type: string
                                  description: The old relationship of the nominee
                                  example: ""
                            newValues:
                              type: object
                              properties:
                                title:
                                  type: string
                                  description: The new title of the nominee
                                  example: "M/s"
                                dob:
                                  type: string
                                  description: The new date of birth of the nominee
                                  example: "29/11/1995"
                                fullname:
                                  type: string
                                  description: The new full name of the nominee
                                  example: "jyotishar"
                                relationship:
                                  type: string
                                  description: The new relationship of the nominee
                                  example: "Brother-in-law"
                        categoryname:
                          type: string
                          description: Category name of the SR
                          example: "Nominee"
                        sub_category_name:
                          type: string
                          description: Sub-category name of the SR
                          example: "Nominee"
                        sr_status:
                          type: string
                          description: Status of the SR
                          example: "Approved"
                        timeEvents:
                          type: array
                          items:
                            type: object
                            properties:
                              event:
                                type: string
                                description: The event name
                                example: "Service Request Raised"
                              time:
                                type: string
                                description: Time of the event
                                example: "07:39 AM"
                              date:
                                type: string
                                format: date
                                description: Date of the event
                                example: "29/11/2024"
        "401":
          description: Unauthorized - Invalid token
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Unauthorized access."
...